requirements.txt
matplotlib>=3.3.0
numpy>=1.19.1
pycairo>=1.19.1
python-igraph>=0.8.2
scipy>=1.5.2

